+++ modules/extensions/src/backend/index.ts
@@ -2,8 +2,8 @@ import * as sdk from 'botpress/sdk'
 import { asyncMiddleware as asyncMw } from 'common/http'
 
 import en from '../translations/en.json'
-import fr from '../translations/fr.json'
 import es from '../translations/es.json'
+import fr from '../translations/fr.json'
 
 const onServerReady = async (bp: typeof sdk) => {
   const asyncMiddleware = asyncMw(bp.logger)