+++ modules/misunderstood/src/backend/index.ts
@@ -1,8 +1,8 @@
 import * as sdk from 'botpress/sdk'
 
 import en from '../translations/en.json'
-import fr from '../translations/fr.json'
 import es from '../translations/es.json'
+import fr from '../translations/fr.json'
 import { FlaggedEvent, FLAG_REASON } from '../types'
 
 import initApi from './api'